+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 * Add DisplayErrorContext to S3 errors
+* Move quilt::_get_obj_attrs to s3_utils::get_attrs_for_key
 
 ## [0.5.8] - 2024-03-25